--- README.md
@@ -1,10 +1,0 @@
-# spectograph
-Simple python tool for creating animated spectrograms
-Requires
-python3 
-numpy
-matplotlib
-...
-
-should be able to run demo with
-python model.py
