+++ model.py
@@ -2,12 +2,13 @@
 This is a small audio/visualization tool. It creates two
 animated graphs and an audio signal.
 """
-
-import numpy as np
 import matplotlib.pyplot as plt
 import matplotlib.animation as animation
 import pyaudio 
-import math
+import numpy as np
+import audio  
+import my_signal
+
 
 LEN_TRACE_VIS = 256	#How much of trace to show
 LEN_TRACE_WIN = 30	#Size of the FFT window
@@ -26,12 +27,8 @@ def initialize( ):
 	Returns them in a dictionary.
 	"""
 	#Signal
-	# the_sig = get_the_sig()
-	the_sig = get_the_collatz(63728127) #Long Collatz seq
+	the_sig = my_signal.get_the_collatz(63728127) #Long Collatz seq
 	the_sig = np.log(the_sig) #Reduce to smaller range
-	# the_sig = the_sig[::-1]
-	#double length
-#	the_sig = np.concatenate( [the_sig, the_sig] )
 
 	max_sig_val = np.amax( the_sig )
 
@@ -64,26 +61,11 @@ def initialize( ):
 	'stream'        : stream
 	}
 
-def get_the_sig():
-	"""
-	Collects the signal.
-	"""
-	return np.fromiter( (x%100 for x in range(1,5000)) , int )
 
-def get_the_collatz(x):
-	seq = []
-	seq = collatz_sequence(seq, x)
-	return np.asarray(seq)
-
-def collatz_sequence(seq, x):
-	seq.append(x)
-	if x == 1:
-		return seq
-	x = x/2 if (x % 2 == 0)  else x*3+1
-	return collatz_sequence(seq, x)
 
 def engine(i,data):	
 	#Trace animation.
+
 	line_t.set_ydata( data['trace_vis'] )
 
 	#Spectrogram animation.
@@ -105,10 +87,9 @@ def engine(i,data):
 
 	#tone
 	data['tone'] = fill / data['max_sig_val'] 
-	# play_tone(data['stream'], data['tone'] * 500 + 30)
-
-	chord = data['spec'][:, data['spec'].shape[1]-1 ][1:]
-	play_chord( data['stream'], chord )
+	audio.play_tone(data['stream'], data['tone'] * 500 + 30)
+	# chord = data['spec'][:, data['spec'].shape[1]-1 ][1:]
+	# audio.play_chord( data['stream'], chord )
 
 	#vis
 	fill = data['the_sig'][LEN_NON_WIN] if ( np.size( data['the_sig'] ) > LEN_NON_WIN ) else 0
@@ -120,11 +101,6 @@ def engine(i,data):
 	#calc fft window sig
 	data['trace_win_fft'] = np.abs( np.fft.rfft ( data['trace_win'] ) )
 
-	#This is only to make visualization easier.
-	#log of magnitudes
-#	data['trace_win_fft'] = np.log( data['trace_win_fft'] )
-	#data['trace_win_fft'] = data['trace_win_fft'] 
-
 	#Fill every element of the spec with one element from the fft.
 	data['spec'] = np.c_[data['spec'], data['trace_win_fft']]	
 
@@ -132,38 +108,6 @@ def printout(data):
 	for key, value in data.items():
 		print(key, value)
 
-def sine(frequency, length, rate):
-	length = int(length * rate)
-	factor = float(frequency) * (math.pi * 2) / rate
-	return np.sin(np.arange(length) * factor)
-
-def sine2(frequency, amp, length, rate):
-	length = int(length * rate)
-	factor = float(frequency) * (math.pi * 2) / rate
-	return amp * np.sin(np.arange(length) * factor)
-
-def play_tone(stream, frequency, length=.1, rate=44100):
-	chunks = []
-	chunks.append(sine(frequency, length, rate))
-	chunk = np.concatenate(chunks) * 0.25
-	stream.write(chunk.astype(np.float32).tostring())
-
-def play_chord(stream, amp, length=.1, rate=44100):
-	freq = np.linspace(200,3000,len(amp))
-	
-	chord = []
-
-	while len(freq) != 0:
-		tmp = []
-		tmp.append(sine2(freq[0], amp[0], length, rate))
-		if chord == []:
-			chord = np.concatenate(tmp) 
-		else:
-			chord += np.concatenate(tmp) 
-		freq = freq[1:]
-		amp = amp[1:]
-
-	stream.write(chord.astype(np.float32).tostring())
 
 #For backend
 data = initialize()
@@ -202,6 +146,5 @@ ax_t.set_axis_bgcolor('grey')
 ax.set_axis_bgcolor('grey')
 
 anim = animation.FuncAnimation(fig, engine, fargs=([data]), interval=1)
-
 plt.show()
